Update raspberrypi.img to contain config.json in the first partition
Back when this module was created, the `config.json` file used to be on
the fourth partition of the image (`RESIN-BOOT`). Nowadays, this file
exists on the first partition.

This module fetches information about the `config.json` location from
the API, so when this change was introduced, this module suddenly
started breaking.

See: https://github.com/resin-io/resin-config-json/pull/4
Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>